[LAS-15] rename: 오타 수정 calander -> calendar, pitness -> fitness #66
