+++ src/components/Common/BottomSheet/BottomSheetList/Calendar/index.tsx
@@ -3,12 +3,12 @@ import * as S from './style';
 import CalendarIcon from '@assets/icon/calendar.svg';
 import BottomArrowIcon from '@assets/icon/bottom-arrow.svg';
 import { useBottomSheet } from '@hooks/common';
-import Calendar from 'react-calendar';
+import ReactCalendar from 'react-calendar';
 import 'react-calendar/dist/Calendar.css';
 
 type ValuePiece = Date | null;
 
-const Calandar = () => {
+const Calendar = () => {
   const { closeBottomSheet } = useBottomSheet();
   const [value, onChange] = useState<ValuePiece | [ValuePiece, ValuePiece]>(new Date());
 
@@ -17,16 +17,16 @@ const Calandar = () => {
   return (
     <S.BottomSheetContent>
       <S.BottomSheetHeader>
-        <span className="select_pitness">
+        <span className="select_fitness">
           <CalendarIcon />
           운동날짜
         </span>
         <BottomArrowIcon />
       </S.BottomSheetHeader>
       <S.Content>
-        <S.CalandarWrapper>
-          <Calendar onChange={onChange} value={value} />
-        </S.CalandarWrapper>
+        <S.CalendarWrapper>
+          <ReactCalendar onChange={onChange} value={value} />
+        </S.CalendarWrapper>
       </S.Content>
       <S.ButtonGroup>
         <S.CancelButton>취소</S.CancelButton>
@@ -36,4 +36,4 @@ const Calandar = () => {
   );
 };
 
-export default Calandar;
+export default Calendar;
