+++ src/components/Common/BottomSheet/BottomSheetList/ExerciseArea/style.tsx
@@ -24,14 +24,14 @@ export const BottomSheetHeader = styled.div`
   border-bottom: 1px solid var(--color-gray3);
   padding-bottom: 15px;
 
-  .select_pitness {
+  .select_fitness {
     font-weight: 700;
     font-size: 16px;
     line-height: 30px;
     letter-spacing: -2.2%;
   }
 
-  .register_pitness {
+  .register_fitness {
     display: flex;
     align-items: center;
     gap: 6px;
