+++ src/components/Common/BottomSheet/BottomSheetList/index.ts
@@ -1,4 +1,4 @@
 export { default as FitnessClub } from './FitnessClub';
 export { default as ExerciseArea } from './ExerciseArea';
 export { default as Gender } from './Gender';
-export { default as Calandar } from './Calandar';
+export { default as Calendar } from './Calendar';
