+++ src/components/Common/BottomSheet/index.tsx
@@ -11,7 +11,7 @@ const BottomSheetContainer = () => {
     FitnessClub: BottomSheet['FitnessClub'],
     ExerciseArea: BottomSheet['ExerciseArea'],
     Gender: BottomSheet['Gender'],
-    Calandar: BottomSheet['Calandar'],
+    Calendar: BottomSheet['Calendar'],
     '': <></>,
   };
   const Content = bottomSheetComp[bottomsheet.name] as () => JSX.Element;
